dev_queue_xmit may actually fail, really
If this function is called to use an interface that is DOWN, it is going to fail. Since IRATI
seems to expect receiving an ARP reply for each ARP requests that it makes, it's important to
know when the request has failed to be sent, otherwise some error will never be handled.

This feels unrelated to the bug entry that was created for this branch but this is what we're
facing: If the IPCM doesn't receive an error when it tries to create the N-1 flow, it will
eventually run out of ports, which locks up pidm.c